refactor: cache フォルダを作成して、そこにあるべきファイルを移動
